requirements.txt
# Core application dependencies
streamlit>=1.32.0
plotly>=5.17.0
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.11.0
requests>=2.31.0
python-dotenv>=1.0.0

# Data sources
yfinance>=0.2.18

# Optional: AI features (requires OpenAI API key)
openai>=1.12.0

# Optional: Additional analysis (not required for core functionality)
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0